power: Fix crash if outlet is None
In get_power_device, the code assumed that if ip_address is None,
then outlet is not equal to None. This is not the case in our
setup and caused a crash of boarfdarm.
This commit adds an extra check to verify that outlet variable is
not None before using it.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>