CheckQosScripts: use opkg list-installed instead of opkg list
opkg list returns the list of available packages which might be
installed. However the purpose of this test is to ensure that
qos-scripts is not installed. Hence, the current implementation
reports fail if the qos-scripts is available even if it is not
installed.
By calling opkg list-installed instead, we make sure that the
test reports a failure only if qos-scripts is installed.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>