Jenkinsfile: support overriding LEDE feed locations
LEDE use a very different standard path to openwrt hence just add it
as another sed command for now. The regex doesn't clash with openwrt
so seems the best comprimise for now.

Note it is designed to replace just the prefix rather than have to
know anything about the paths used further into the tree.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>